remove cprof artifacts, remove http server, move logger and config module to app module
