--- cpu.prof
@@ -1,1 +1,0 @@
-Could not enable CPU profiling: cpu profiling already in use
